REFACTOR: Switch all js files to kebab case
